Fix compiler warnings

+++ src/pb/Grid.cc
@@ -130,9 +130,9 @@ Grid::Grid(const Grid& my_grid, const short nghosts)
 
     if (nghosts == -1)
     {
-        assert(my_grid.ghost_pt_ <= dim_[0]);
-        assert(my_grid.ghost_pt_ <= dim_[1]);
-        assert(my_grid.ghost_pt_ <= dim_[2]);
+        assert(static_cast<unsigned int>(my_grid.ghost_pt_) <= dim_[0]);
+        assert(static_cast<unsigned int>(my_grid.ghost_pt_) <= dim_[1]);
+        assert(static_cast<unsigned int>(my_grid.ghost_pt_) <= dim_[2]);
 
         ghost_pt_ = my_grid.ghost_pt_; // default
         inc_[0]   = my_grid.inc_[0];
@@ -142,9 +142,9 @@ Grid::Grid(const Grid& my_grid, const short nghosts)
     }
     else
     {
-        assert(nghosts <= dim_[0]);
-        assert(nghosts <= dim_[1]);
-        assert(nghosts <= dim_[2]);
+        assert(static_cast<unsigned int>(nghosts) <= dim_[0]);
+        assert(static_cast<unsigned int>(nghosts) <= dim_[1]);
+        assert(static_cast<unsigned int>(nghosts) <= dim_[2]);
 
         ghost_pt_ = nghosts;
         inc_[0]   = (dim_[1] + 2 * ghost_pt_) * (dim_[2] + 2 * ghost_pt_);
